refactor: code quality improvements (PR F) (#51)
Implement comprehensive code quality improvements:

F1: Replace all 8 DynamicJsonDocument with StaticJsonDocument + overflow checks
    in ControllerWebServer.cpp. Each JSON endpoint now validates buffer capacity
    and returns HTTP 500 on overflow instead of silent truncation.

F2: Limit CAN frame processing to MAX_CAN_FRAMES_PER_TICK (10 frames per tick)
    in main.cpp checkCanbus(). Prevents loop starvation when CAN bus is flooded.

F3: Extract CAN payload parsers as static free functions in unnamed namespaces:
    - HobbywingCan: 6 static parser functions for telemetry field extraction
    - TmotorCan: 5 static helpers (parseUint16LE, parseUint32LE, etc.) for
      little-endian multi-byte field parsing, improving code modularity

F4: Replace magic numbers with named constants:
    - HobbywingCan: STATUS_MSG1_FRAME_LENGTH, TEMP_OFFSET, DIRECTION_MASK, etc.
    - TmotorCan: ESC_STATUS_*_OFFSET, PUSHCAN_*, STATUS5_*, TRANSFER_ID_MASK,
      MAX_THROTTLE, PARAMCFG_SIGNATURE, CAN_PAYLOAD_SIZE

F5: Convert Steinhart-Hart temperature calculation from double/log() to
    float/logf() in Temperature.cpp. Also fixed memcpy→memmove for overlapping
    buffer regions.

All 3 build targets verified (hobbywing, tmotor, xag) with no warnings.

+++ src/Temperature/Temperature.cpp
@@ -46,17 +46,17 @@ void Temperature::readTemperature() {
   }
 
   // Voltage at the divider point: ReadFn returns 0-4095, scale by adcVoltageRef
-  double v = (adcVoltageRef * (double)sum) / (samples * 4095.0);
+  float v = (adcVoltageRef * (float)sum) / (samples * 4095.0f);
 
   // Solving for rt: rt = (v * R) / (VCC - v)
   // VCC is 3.3V (sensor power supply)
-  const double vcc = 3.3;
-  double rt = (v * R) / (vcc - v);
+  const float vcc = 3.3f;
+  float rt = (v * R) / (vcc - v);
 
   // Steinhart-Hart equation using Beta coefficient:
   // 1/T = 1/T0 + (1/B) * ln(Rt/R0)
   // T = 1 / (1/T0 + (1/B) * ln(Rt/R0))
-  double invT = (1.0 / t0) + (1.0 / beta) * log(rt / r0);
-  double tempK = 1.0 / invT;
-  temperature = tempK - 273.15;
+  float invT = (1.0f / t0) + (1.0f / beta) * logf(rt / r0);
+  float tempK = 1.0f / invT;
+  temperature = tempK - 273.15f;
 }
